update-deviceController

+++ app/Http/Controllers/Api/user/DeviceController.php
@@ -10,6 +10,7 @@ use App\Models\Device;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
 use App\Http\Helpers\ImageUploader;
+use App\Models\Room;
 
 class DeviceController extends Controller
 {
@@ -17,60 +18,98 @@ class DeviceController extends Controller
     use ApiResponse;
 
 
-    public function index(){
+    public function index($room_id){
+        $room=Room::where('id',$room_id)->first();
+
+        if (!$room) {
+            return $this->apiResponse('', "room not found");
+        }
+
+        if($room->user_id!==auth('user')->user()->id){
+
+            return $this->apiResponse('', "not have permissions");
+
+        }
+
+        $devices=$room->devices;
+        $data = DeviceResource::collection($devices);
+        return $this->apiResponse($data,"return data successfully");
+    }
+
+    public function getAllDevices()
+    {
         $devices=Device::where('user_id',auth()->user()->id)->get();
         $data = DeviceResource::collection($devices);
         return $this->apiResponse($data,"return data successfully");
     }
 
-    public function allByRoom($room_id)
-    {
-        $devices=Device::where('user_id',auth()->user()->id)
-                        ->where('room_id',$room_id)->get();
-        $data = DeviceResource::collection($devices);
-        return $this->apiResponse($data,"return data successfully");
-    }
 
-    
-    
-    public function store(DeviceRequest $request)
+
+    public function store(DeviceRequest $request, $room_id)
     {
-        $data=$request->validated();
-        
-        $data['user_id']=auth()->user()->id;
-        
+        $room=Room::where('id',$room_id)->first();
+
+        if (!$room) {
+            return $this->apiResponse('', "room not found");
+        }
+
+        if($room->user_id!==auth('user')->user()->id){
+
+            return $this->apiResponse('', "not have permissions");
+
+        }
+
+        $data['name']=$request->name;
+        $data['room_id']=$room_id;
+        $data['user_id']=auth('user')->user()->id;
+
         $Device=Device::create($data);
 
         $data = DeviceResource::make($Device);
-        
+
         return $this->apiResponse($data,"stored data successfully");
     }
 
-    
-    
-    public function show($id)
+
+
+    public function show($device_id)
     {
-        $device = Device::where('user_id',auth()->user()->id)
-                        ->where('id',$id)->first();
+
+
+        $device = Device::where('id',$device_id)->first();
+
         if(!$device){
-            return $this->apiResponse('',"Device not found"); 
+            return $this->apiResponse('',"Device not found");
         }
-        
+
+        if($device->user_id!==auth('user')->user()->id){
+
+            return $this->apiResponse('', "not have permissions");
+
+        }
+
         $data = DeviceResource::make($device);
         return $this->apiResponse($data,"return data successfully");
-        
+
 
 
     }
 
- 
-    
-    public function update(DeviceRequest $request, $id)
+
+
+    public function update(DeviceRequest $request, $device_id)
     {
 
-        $device = Device::find($id);
+        $device = Device::where('id',$device_id)->first();
+
         if(!$device){
-            return $this->apiResponse('',"Device not found"); 
+            return $this->apiResponse('',"Device not found");
+        }
+
+        if($device->user_id!==auth('user')->user()->id){
+
+            return $this->apiResponse('', "not have permissions");
+
         }
 
         $data=$request->validated();
@@ -83,18 +122,25 @@ class DeviceController extends Controller
 
     }
 
-   
-    
-    public function destroy($id)
+
+
+    public function destroy($device_id)
     {
-        $device = Device::find($id);
+        $device = Device::where('id',$device_id)->first();
+
         if(!$device){
-            return $this->apiResponse('',"Device not found"); 
+            return $this->apiResponse('',"Device not found");
         }
-        
+
+        if($device->user_id!==auth('user')->user()->id){
+
+            return $this->apiResponse('', "not have permissions");
+
+        }
+
         $device->delete();
         $data = DeviceResource::make($device);
         return $this->apiResponse($data,"deleted data successfully");
-        
+
    }
 }
